% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_from_log.R
\name{code_from_log}
\alias{code_from_log}
\title{Create R Script File from R Log File}
\usage{
code_from_log(log_file, code_file)
}
\arguments{
\item{log_file}{character string. Path and name of the log file produced with
\code{produce_log_file()}}

\item{code_file}{character string. Path and name of the resulting R Script
file}
}
\description{
After running \code{produce_log_file()} this function creates an R Script
from the resulting log file. Accordingly, it deletes all information on
results, errors and other other messages.
}
