% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_file.R
\name{current_file}
\alias{current_file}
\title{https://stackoverflow.com/a/36777602
current script file (in full path)}
\usage{
current_file()
}
\description{
current script file (in full path)
}
\examples{
works with Rscript, source() or in RStudio Run selection, RStudio Console
}
