% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_log_file.R
\name{produce_log_file}
\alias{produce_log_file}
\title{Evaluate R Source File and Write It to Log File}
\usage{
produce_log_file(source_file, log_file)
}
\arguments{
\item{source_file}{character string. Location and name of R source file that
should be evaluated and logged. Note that the 3 last lines of this source
file are reserved for this function. Accordingly, this function should be
placed there. They get deleted on the fly after execution. Therefore, they
won't appear in the log file.}

\item{log_file}{character string. Location and name of log file to which logs
are written}
}
\description{
Mimics Stata's log file function. The function expects to be placed at the
bottom (last 3 lines) of the R script that should be logged.
}
